requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
paramiko==4.0.0
pycparser==3.0
PyNaCl==1.6.2
python-dotenv==1.0.1
redis==7.2.0
requests==2.32.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
websockify==0.12.0
Werkzeug==3.1.6
